Rename and move Error module (helpers)
